feat: innerClass 로 particiPantResult 이동
